tests/wpt/web-platform-tests/css/css-box/margin-trim/computed-margin-values/flexbox-column-multi-line-block-start.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-flex">
<meta name="assert" content="block-start trimmed margins should be reflected in computed style">
<style>
flexbox {
    display: flex;
    flex-direction: column;
    flex-wrap: wrap;
    width: 10px;
    height: 110px;
    margin-trim: block-start;
}
item {
    display: block;
    background-color: green;
    width: 50px;
    height: 50px;
    margin-block-start: 10px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
</head>
<body onload="checkLayout('flexbox > item')">
<div id="target">
<flexbox>
    <item data-offset-y="8" data-expected-margin-top="0"></item>
    <item data-offset-y="68" data-expected-margin-top="10"></item>
    <item data-offset-y="8" data-expected-margin-top="0"></item>
    <item data-offset-y="68" data-expected-margin-top="10"></item>
</flexbox>
</div>
</body>
</html>
