third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-multi-line-block-start-expected.txt

This is a testharness.js-based test.
FAIL flexbox > item 1 assert_equals: 
<item data-offset-y="8" data-expected-margin-top="0"></item>
offsetTop expected 8 but got 18
FAIL flexbox > item 2 assert_equals: 
<item data-offset-y="68" data-expected-margin-top="10"></item>
offsetTop expected 68 but got 18
FAIL flexbox > item 3 assert_equals: 
<item data-offset-y="8" data-expected-margin-top="0"></item>
offsetTop expected 8 but got 18
FAIL flexbox > item 4 assert_equals: 
<item data-offset-y="68" data-expected-margin-top="10"></item>
offsetTop expected 68 but got 18
Harness: the test ran to completion.
